refactor(Sections.java) : Validator 제거하고 검증로직 Sections 내부로 이동
