+++ src/main/java/wooteco/subway/line/domain/section/Sections.java
@@ -1,10 +1,5 @@
 package wooteco.subway.line.domain.section;
 
-import wooteco.subway.line.domain.section.validate.add.SectionAddValidator;
-import wooteco.subway.line.domain.section.validate.add.SectionAddValidatorImpl;
-import wooteco.subway.line.domain.section.validate.delete.SectionDeleteValidator;
-import wooteco.subway.line.domain.section.validate.delete.SectionDeleteValidatorImpl;
-
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -13,15 +8,12 @@ public class Sections {
     private static final Sections EMPTY = new Sections(Collections.emptyList());
 
     private static final int END_POINT_COUNT = 1;
+    private static final int DELETE_LIMIT_SIZE = 1;
 
     private final List<Section> sections;
-    private final SectionAddValidator addValidator;
-    private final SectionDeleteValidator deleteValidator;
 
     public Sections(List<Section> sections) {
         this.sections = sections;
-        this.addValidator = new SectionAddValidatorImpl();
-        this.deleteValidator = new SectionDeleteValidatorImpl();
     }
 
     public static Sections empty() {
@@ -29,12 +21,31 @@ public class Sections {
     }
 
     public void add(Section section) {
-        addValidator.validatePossibleToAdd(this, section);
+        validatePossibleToAdd(this, section);
         if (isAddToEndPoint(section)) {
             sections.add(section);
             return;
         }
         addToBetween(section);
+    }
+
+    private void validatePossibleToAdd(Sections sections, Section newSection) {
+        boolean existUpStation = sections.isExistStationId(newSection.getUpStationId());
+        boolean existDownStation = sections.isExistStationId(newSection.getDownStationId());
+        validateNotExistSectionOfStation(existUpStation, existDownStation);
+        validateAlreadyExistSectionOfStation(existUpStation, existDownStation);
+    }
+
+    private void validateNotExistSectionOfStation(boolean existUpStation, boolean existDownStation) {
+        if (!existUpStation && !existDownStation) {
+            throw new IllegalArgumentException("연결할 수 있는 역이 구간내에 없습니다.");
+        }
+    }
+
+    private void validateAlreadyExistSectionOfStation(boolean existUpStation, boolean existDownStation) {
+        if (existUpStation && existDownStation) {
+            throw new IllegalArgumentException("상행역과 하행역이 이미 존재합니다.");
+        }
     }
 
     private void addToBetween(Section newSection) {
@@ -76,12 +87,30 @@ public class Sections {
     }
 
     public void delete(Long stationId) {
-        deleteValidator.validateDeleteSection(this, stationId);
+        validateDeleteSection(this, stationId);
         if (isDeleteToEndPoint(stationId)) {
             deleteEndPoint(stationId);
             return;
         }
         deleteSectionOfStation(stationId);
+    }
+
+    private void validateDeleteSection(Sections sections, Long stationId) {
+        validateExistStationId(sections, stationId);
+        validateDeleteSize(sections);
+    }
+
+    private void validateExistStationId(Sections sections, Long stationId) {
+        if (sections.isExistStationId(stationId)) {
+            return;
+        }
+        throw new IllegalArgumentException("삭제하려는 역을 포함하는 구간이 존재하지 않습니다.");
+    }
+
+    private void validateDeleteSize(Sections sections) {
+        if (sections.toList().size() <= DELETE_LIMIT_SIZE) {
+            throw new IllegalStateException("구간이 하나 이하일 때는 삭제할 수 없습니다.");
+        }
     }
 
     private void deleteSectionOfStation(Long stationId) {
@@ -111,7 +140,7 @@ public class Sections {
         return this.stationIds().contains(stationId);
     }
 
-    public List<Long> stationIds() {
+    private List<Long> stationIds() {
         return sections.stream()
                 .flatMap(section -> Stream.of(section.getUpStationId(), section.getDownStationId()))
                 .distinct()
@@ -121,19 +150,25 @@ public class Sections {
     private Sections sortSection() {
         Deque<Section> sectionDeque = new ArrayDeque<>();
         sectionDeque.add(sections.get(0));
+        sortUpSections(sectionDeque);
+        sortDownSections(sectionDeque);
+        return new Sections(new ArrayList<>(sectionDeque));
+    }
 
-        Map<Long, Section> upStationIdMap = sections.stream()
+    private void sortUpSections(Deque<Section> sectionDeque) {
+        Map<Long, Section> upStationIdMap = this.sections.stream()
                 .collect(Collectors.toMap(Section::getUpStationId, section -> section));
         while (upStationIdMap.containsKey(sectionDeque.peekLast().getDownStationId())) {
             sectionDeque.addLast(upStationIdMap.get(sectionDeque.peekLast().getDownStationId()));
         }
+    }
 
-        Map<Long, Section> downStationIdMap = sections.stream()
+    private void sortDownSections(Deque<Section> sectionDeque) {
+        Map<Long, Section> downStationIdMap = this.sections.stream()
                 .collect(Collectors.toMap(Section::getDownStationId, section -> section));
         while (downStationIdMap.containsKey(sectionDeque.peekFirst().getUpStationId())) {
             sectionDeque.addFirst(downStationIdMap.get(sectionDeque.peekFirst().getUpStationId()));
         }
-        return new Sections(new ArrayList<>(sectionDeque));
     }
 
     public List<Long> sortSectionsId() {
